Separating out bezel notification parameters
This way the system bezel notifications AND the bezel notification toasts can both use the same parameters.
Also means shuffling logic around, and a few other such chores which shoulda been done awhile ago
