+++ Sources/BlueToast/Legacy/SystemBezelNotification.swift
@@ -212,136 +212,7 @@ public extension SystemBezelNotification {
 
 
 
-/// A set of parameters used to configure and present a bezel notification
 public extension SystemBezelNotification {
-    struct Parameters {
-        
-        public static let defaultLocation: Location = .normal
-        public static let defaultSize: Size = .normal
-        public static let defaultTimeToLive: TimeToLive = .short
-        
-        public static let defaultFadeInAnimationDuration: TimeInterval = 0
-        public static let defaultFadeOutAnimationDuration: TimeInterval = 0.25
-        
-        public static let defaultCornerRadius: CGFloat = 18
-        public static let defaultBackgroundTint = NSColor.clear
-        public static let defaultMessageLabelBaselineOffsetFromBottomOfBezel: CGFloat = 20
-        public static let defaultMessageLabelFontSize: CGFloat = 18
-        public static let defaultMessageLabelFont = NSFont.systemFont(ofSize: defaultMessageLabelFontSize)
-        public static let defaultMessageLabelColor = NSColor.labelColor
-        
-        
-        // MARK: Basics
-        
-        /// The text to show in the bezel notification's message area
-        var messageText: String
-        
-        /// The icon to show in the bezel notification's icon area
-        var icon: NSImage?
-        
-        
-        // MARK: Presentation
-        
-        /// The location on the screen at which to display the bezel notification
-        var location: Location
-        
-        /// The size of the bezel notification
-        var size: Size
-        
-        /// The number of seconds to display the bezel notification on the screen
-        var timeToLive: TimeToLive
-        
-        
-        // MARK: Animations
-        
-        /// The number of seconds that it takes to fade in the bezel notification
-        var fadeInAnimationDuration: TimeInterval
-        
-        /// The number of seconds that it takes to fade out the bezel notification
-        var fadeOutAnimationDuration: TimeInterval
-        
-        
-        // MARK: Drawing
-        
-        /// The radius of the bezel notification's corners, in points
-        var cornerRadius: CGFloat
-        
-        /// The tint of the bezel notification's background
-        var backgroundTint: NSColor
-        
-        /// The distance from the bottom of the bezel notification's bottom at which the baseline of the message label sits
-        var messageLabelBaselineOffsetFromBottomOfBezel: CGFloat
-        
-        /// The font used for the message label
-        var messageLabelFont: NSFont
-        
-        /// The text color of the message label
-        var messageLabelColor: NSColor
-        
-        
-        public init(messageText: String,
-                    icon: NSImage? = nil,
-                    
-                    location: Location = defaultLocation,
-                    size: Size = defaultSize,
-                    timeToLive: TimeToLive = defaultTimeToLive,
-                    
-                    fadeInAnimationDuration: TimeInterval = defaultFadeInAnimationDuration,
-                    fadeOutAnimationDuration: TimeInterval = defaultFadeOutAnimationDuration,
-                    
-                    cornerRadius: CGFloat = defaultCornerRadius,
-                    backgroundTint: NSColor = defaultBackgroundTint,
-                    messageLabelBaselineOffsetFromBottomOfBezel: CGFloat = defaultMessageLabelBaselineOffsetFromBottomOfBezel,
-                    messageLabelFont: NSFont = defaultMessageLabelFont,
-                    messageLabelColor: NSColor = defaultMessageLabelColor
-        ) {
-            self.messageText = messageText
-            self.icon = icon
-            
-            self.location = location
-            self.size = size
-            self.timeToLive = timeToLive
-            
-            self.fadeInAnimationDuration = fadeInAnimationDuration
-            self.fadeOutAnimationDuration = fadeOutAnimationDuration
-            
-            self.cornerRadius = cornerRadius
-            self.backgroundTint = backgroundTint.withAlphaComponent(backgroundTint.alphaComponent * 0.15)
-            self.messageLabelBaselineOffsetFromBottomOfBezel = messageLabelBaselineOffsetFromBottomOfBezel
-            self.messageLabelFont = messageLabelFont
-            self.messageLabelColor = messageLabelColor
-        }
-    }
-    
-    
-    
-    /// How long a bezel notification should stay on screen
-    enum TimeToLive {
-        
-        /// Bezel is shown for just a couple seconds
-        case short
-        
-        /// Bezel if shown for several seconds
-        case long
-        
-        /// Bezel is never hidden
-        case forever
-        
-        /// Bezel is shown for an exact number of seconds
-        case exactly(seconds: TimeInterval)
-        
-        
-        var inSeconds : TimeInterval {
-            switch self {
-            case .short: return 2
-            case .long: return 6
-            case .forever: return .infinity
-            case .exactly(let seconds): return seconds
-            }
-        }
-    }
-    
-    
     
     /// The window used to present a bezel notification.
     /// If you _really_ need minute control, you may use this.
@@ -350,7 +221,10 @@ public extension SystemBezelNotification {
         private lazy var bezelContentView: ContentView = {
             let bezelContentView = ContentView(parameters: self.parameters)
             bezelContentView.wantsLayer = true
-            bezelContentView.layer?.backgroundColor = parameters.backgroundTint.cgColor
+            let inputBackgroundTint = parameters.backgroundTint
+            let backgroundTint = inputBackgroundTint.withAlphaComponent(inputBackgroundTint.alphaComponent * 0.15)
+
+            bezelContentView.layer?.backgroundColor = backgroundTint.cgColor
             return bezelContentView
         }()
         
@@ -365,8 +239,10 @@ public extension SystemBezelNotification {
         public init(parameters: Parameters) {
             
             self.parameters = parameters
-            
-            let contentRect = parameters.location.bezelWindowContentRect(atSize: parameters.size)
+            let screen = parameters.location.screen
+            let contentRect = parameters.location.bezelWindowContentRect(
+                in: screen ?? CGRect(origin: CGPoint(x: 48, y: 48), size: size.cgSize),
+                atSize: parameters.size)
             
             super.init(contentRect: contentRect,
                        styleMask: bezelStyleMask,
@@ -488,23 +364,16 @@ public extension SystemBezelNotification {
 }
 
 
-extension SystemBezelNotification.TimeToLive: Hashable {
-    public func hash(into hasher: inout Hasher) {
+
+internal extension BezelNotificationParameters.Location {
+    var screen: NSScreen? {
         switch self {
-        case .short:
-            hasher.combine("\(Self.self).__PREDEFINED__.short")
-            
-        case .long:
-            hasher.combine("\(Self.self).__PREDEFINED__.long")
-            
-        case .forever:
-            hasher.combine("\(Self.self).__PREDEFINED__.forever")
-            
-        case .exactly(let seconds):
-            hasher.combine("\(Self.self).exactly(seconds: \(seconds))")
+        case .normal:
+            return .main ?? NSScreen.screens.first
         }
     }
 }
+
 
 
 private extension String {
@@ -529,76 +398,7 @@ private extension String {
 
 
 public extension SystemBezelNotification {
-    /// The semantic size of a bezel notification
-    enum Size {
-        case normal
-    }
+    /// A set of parameters used to configure and present a bezel notification
+    typealias Parameters = BezelNotificationParameters
 }
-
-
-
-internal extension SystemBezelNotification.Size {
-    
-    private var width: CGFloat {
-        switch self {
-        case .normal:
-            return 200
-        }
-    }
-    
-    
-    private var height: CGFloat {
-        switch self {
-        case .normal:
-            return 200
-        }
-    }
-    
-    
-    var cgSize: CGSize {
-        CGSize(width: width, height: height)
-    }
-}
-
-
-
-public extension SystemBezelNotification {
-    /// The semantic location of a bezel notification
-    enum Location {
-        case normal
-    }
-}
-
-
-
-internal extension SystemBezelNotification.Location {
-    func bezelWindowContentRect(atSize size: SystemBezelNotification.Size) -> NSRect {
-        switch self {
-        case .normal:
-            return screen?.lowerCenterRect(ofSize: size.cgSize) ?? NSRect(origin: NSPoint(x: 48, y: 48), size: size.cgSize)
-        }
-    }
-    
-    
-    var screen: NSScreen? {
-        switch self {
-        case .normal:
-            return .main ?? NSScreen.screens.first
-        }
-    }
-}
-
-
-
-private extension NSScreen {
-    
-    private static let lowerCenterRectBottomOffset: CGFloat = 140
-    
-    func lowerCenterRect(ofSize size: NSSize) -> NSRect {
-        return NSRect(origin: NSPoint(x: self.frame.midX - (size.width / 2),
-                                      y: self.frame.minY + NSScreen.lowerCenterRectBottomOffset),
-                      size: size)
-    }
-}
-
 #endif
